+++ Code/Engine/Foundation/Threading/Implementation/Task.cpp
@@ -14,7 +14,7 @@ void ezTask::Reset()
   m_bUsesMultiplicity = m_uiMultiplicity > 0;
 }
 
-void ezTask::ConfigureTask(const char* szTaskName, ezTaskNesting nestingMode, OnTaskFinishedCallback Callback /*= OnTaskFinishedCallback()*/)
+void ezTask::ConfigureTask(const char* szTaskName, ezTaskNesting nestingMode, ezOnTaskFinishedCallback Callback /*= ezOnTaskFinishedCallback()*/)
 {
   EZ_ASSERT_DEV(IsTaskFinished(), "This function must be called before the task is started.");
