+++ Code/Engine/Foundation/Threading/Implementation/Task.h
@@ -36,7 +36,7 @@ public:
   /// \param Callback
   /// A callback to execute when the task is finished (or canceled).
   /// The most common use case for this is to deallocate the task at that time.
-  void ConfigureTask(const char* szTaskName, ezTaskNesting nestingMode, OnTaskFinishedCallback Callback = OnTaskFinishedCallback()); // [tested]
+  void ConfigureTask(const char* szTaskName, ezTaskNesting nestingMode, ezOnTaskFinishedCallback Callback = ezOnTaskFinishedCallback()); // [tested]
 
   /// \brief Changes the multiplicity of this task.
   ///
@@ -114,7 +114,7 @@ private:
   ezTaskNesting m_NestingMode = ezTaskNesting::Maybe;
 
   /// \brief Optional callback to be fired when the task has finished or was canceled.
-  OnTaskFinishedCallback m_OnTaskFinished;
+  ezOnTaskFinishedCallback m_OnTaskFinished;
 
   /// \brief The parent group to which this task belongs.
   ezTaskGroupID m_BelongsToGroup;
