+++ Code/Engine/Foundation/Threading/Implementation/TaskGroups.cpp
@@ -7,7 +7,7 @@
 #include <Foundation/Threading/Lock.h>
 #include <Foundation/Threading/TaskSystem.h>
 
-ezTaskGroupID ezTaskSystem::CreateTaskGroup(ezTaskPriority::Enum Priority, OnTaskGroupFinishedCallback callback)
+ezTaskGroupID ezTaskSystem::CreateTaskGroup(ezTaskPriority::Enum Priority, ezOnTaskGroupFinishedCallback callback)
 {
   EZ_LOCK(s_TaskSystemMutex);
 
@@ -218,59 +218,6 @@ void ezTaskSystem::ScheduleGroupTasks(ezTaskGroup* pGroup, bool bHighPriority)
         // nothing to do for these enum values
         break;
     }
-  }
-}
-
-ezResult ezTaskSystem::WakeUpThreadIfIdle(ezWorkerThreadType::Enum type, ezUInt32 threadIdx)
-{
-  if (s_WorkerThreads[type][threadIdx]->WakeUpIfIdle().Succeeded())
-  {
-    // the thread index must be different, if it is the same, it must be an entirely different worker thread type
-    EZ_ASSERT_DEV(threadIdx != tl_TaskWorkerInfo.m_iWorkerIndex || type != tl_TaskWorkerInfo.m_WorkerType, "Calling thread was in idle state itself.");
-
-    return EZ_SUCCESS;
-  }
-
-  return EZ_FAILURE;
-}
-
-ezInt32 ezTaskSystem::CalcActivatableThreads(ezWorkerThreadType::Enum type)
-{
-  const ezUInt32 uiTotal = s_iNumWorkerThreads[type];
-  const ezUInt32 uiBlocked = s_BlockedWorkerThreads[type];
-  const ezUInt32 uiIdle = s_IdleWorkerThreads[type];
-  const ezUInt32 uiMaxActive = s_MaxWorkerThreadsToUse[type];
-
-  EZ_ASSERT_DEV(uiBlocked <= uiTotal, "Incorrect worker thread statistics");
-  EZ_ASSERT_DEV(uiIdle <= uiTotal, "Incorrect worker thread statistics");
-  EZ_ASSERT_DEV(uiIdle + uiBlocked <= uiTotal, "Incorrect worker thread statistics");
-
-  const ezUInt32 uiActive = uiTotal - uiBlocked - uiIdle;
-
-  const ezInt32 iCanActivate = (ezInt32)uiMaxActive - (ezInt32)uiActive;
-
-  return iCanActivate;
-}
-
-void ezTaskSystem::WakeUpThreads(ezWorkerThreadType::Enum type, ezUInt32 uiNumThreadsToWakeUp)
-{
-  const ezInt32 iCanActivate = CalcActivatableThreads(type);
-  ezInt32 iShouldActivate = ezMath::Min<ezInt32>(iCanActivate, uiNumThreadsToWakeUp);
-
-  const ezUInt32 uiTotal = s_iNumWorkerThreads[type];
-
-  for (ezUInt32 threadIdx = 0; threadIdx < uiTotal && iShouldActivate > 0; ++threadIdx)
-  {
-    if (WakeUpThreadIfIdle(type, threadIdx).Succeeded())
-    {
-      --iShouldActivate;
-    }
-  }
-
-  if (iShouldActivate > 0)
-  {
-    // the new threads will start not-idle and take on some work
-    AllocateThreads(type, iShouldActivate);
   }
 }
